add a custom hook for Complex lookup
