+++ src/common/constants/complexLookup.constants.ts
@@ -57,4 +57,7 @@ export const COMPLEX_LOOKUPS_LINKED_FIELDS_MAPPING = {
   },
 };
 
-export const EMPTY_LINKED_DROPDOWN_OPTION_SUFFIX = 'empty'
+export const EMPTY_LINKED_DROPDOWN_OPTION_SUFFIX = 'empty';
+export const VALUE_DIVIDER = ' ,';
+export const __MOCK_URI_CHANGE_WHEN_IMPLEMENTING = '__MOCK_URI_CHANGE_WHEN_IMPLEMENTING';
+
