+++ src/components/ComplexLookupField/ComplexLookupField.tsx
@@ -1,103 +1,32 @@
-import { ChangeEvent, FC, useCallback, useContext, useState } from 'react';
-import { useRecoilState, useRecoilValue } from 'recoil';
+import { FC } from 'react';
 import { FormattedMessage } from 'react-intl';
 import classNames from 'classnames';
-import { useModalControls } from '@common/hooks/useModalControls';
 import { IS_EMBEDDED_MODE } from '@common/constants/build.constants';
-import { COMPLEX_LOOKUPS_CONFIG } from '@common/constants/complexLookup.constants';
-import {
-  generateEmptyValueUuid,
-  getLinkedField,
-  getUpdatedSelectedEntries,
-  updateLinkedFieldValue,
-} from '@common/helpers/complexLookup.helper';
+import { COMPLEX_LOOKUPS_CONFIG, VALUE_DIVIDER } from '@common/constants/complexLookup.constants';
+import { useComplexLookup } from '@common/hooks/useComplexLookup';
 import { Input } from '@components/Input';
-import { ServicesContext } from '@src/contexts';
-import state from '@state';
 import CloseIcon from '@src/assets/times-16.svg?react';
 import { ModalComplexLookup } from './ModalComplexLookup';
 import './ComplexLookupField.scss';
 
 interface Props {
-  uuid: string;
   entry: SchemaEntry;
   value?: UserValueContents[];
   onChange: (uuid: string, contents: Array<UserValueContents>) => void;
 }
 
-const __MOCK_URI_CHANGE_WHEN_IMPLEMENTING = '__MOCK_URI_CHANGE_WHEN_IMPLEMENTING';
-const VALUE_DIVIDER = ', ';
-
-export const ComplexLookupField: FC<Props> = ({ value = undefined, uuid, entry, onChange }) => {
-  const { selectedEntriesService } = useContext(ServicesContext);
-  const [selectedEntries, setSelectedEntries] = useRecoilState(state.config.selectedEntries);
-  const schema = useRecoilValue(state.config.schema);
-  const [localValue, setLocalValue] = useState<UserValueContents[]>(value || []);
-  const { isModalOpen, setIsModalOpen, openModal } = useModalControls();
-  const { layout, linkedEntry } = entry;
+export const ComplexLookupField: FC<Props> = ({ value = undefined, entry, onChange }) => {
+  const { layout } = entry;
   const lookupConfig = COMPLEX_LOOKUPS_CONFIG[layout?.api as string];
-  const linkedField = getLinkedField({ schema, linkedEntry });
+  const buttonConfigLabel = lookupConfig?.labels?.button;
 
-  const handleOnChangeBase = ({ target: { value } }: ChangeEvent<HTMLInputElement>) => {
-    const newValue = {
-      label: value,
-      meta: {
-        uri: __MOCK_URI_CHANGE_WHEN_IMPLEMENTING,
-      },
-    };
-
-    onChange(uuid, [newValue]);
-    setLocalValue(prevValue => [...prevValue, newValue]);
-  };
-
-  const handleDelete = (id?: string) => {
-    onChange(uuid, []);
-    setLocalValue(prevValue => prevValue.filter(({ id: prevId }) => prevId !== id));
-
-    if (!(linkedField && selectedEntriesService)) return;
-
-    const updatedSelectedEntries = getUpdatedSelectedEntries({
-      selectedEntriesService,
-      selectedEntries,
-      linkedFieldChildren: linkedField.children,
-      newValue: generateEmptyValueUuid(linkedField.uuid),
+  const { localValue, isModalOpen, openModal, closeModal, handleDelete, handleAssign, handleOnChangeBase } =
+    useComplexLookup({
+      entry,
+      value,
+      lookupConfig,
+      onChange,
     });
-
-    setSelectedEntries(updatedSelectedEntries);
-  };
-
-  const closeModal = useCallback(() => {
-    setIsModalOpen(false);
-  }, []);
-
-  const onAssign = useCallback(({ id, title, linkedFieldValue }: ComplexLookupAssignRecordDTO) => {
-    const newValue = {
-      label: title,
-      meta: {
-        id,
-        uri: '',
-      },
-    };
-
-    onChange(uuid, [newValue]);
-    setLocalValue([newValue]);
-
-    // Has an associated dependent subfield.
-    // For now we assume that the associated field's type is Dropdown only
-    if (linkedEntry?.secondary) {
-      const updatedValue = updateLinkedFieldValue({ schema, linkedField, linkedFieldValue, lookupConfig });
-      const updatedSelectedEntries = getUpdatedSelectedEntries({
-        selectedEntriesService,
-        selectedEntries,
-        linkedFieldChildren: linkedField?.children,
-        newValue: updatedValue?.uuid,
-      });
-
-      setSelectedEntries(updatedSelectedEntries);
-    }
-
-    closeModal();
-  }, []);
 
   return (
     <>
@@ -131,15 +60,13 @@ export const ComplexLookupField: FC<Props> = ({ value = undefined, uuid, entry, 
           )}
 
           <button role="button" className="complex-lookup-select-button button-passive" onClick={openModal}>
-            <FormattedMessage
-              id={localValue?.length ? lookupConfig?.labels?.button.change : lookupConfig?.labels?.button.base}
-            />
+            <FormattedMessage id={localValue?.length ? buttonConfigLabel?.change : buttonConfigLabel?.base} />
           </button>
 
           <ModalComplexLookup
             isOpen={isModalOpen}
             onClose={closeModal}
-            onAssign={onAssign}
+            onAssign={handleAssign}
             assignEntityName={layout.api}
             baseLabelType={layout.baseLabelType}
           />
